+++ coupon_limit/models/__init__.py
@@ -1,0 +1,1 @@
+from . import coupon_rule
