+++ coupon_limit/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module allows to configure a limit on the times a promotion can be applied. It's a
+technical base to be extended with top level functionality (sales, pos).
