+++ coupon_limit/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_coupon_limit
