+++ sale_coupon_limit/models/__init__.py
@@ -1,3 +1,3 @@
-from . import sale_coupon_rule
-from . import sale_coupon_program
-from . import sale_coupon
+from . import coupon_coupon
+from . import coupon_program
+from . import coupon_rule
