+++ setup/coupon_limit/odoo/addons/coupon_limit
@@ -1,0 +1,1 @@
+../../../../coupon_limit
